Move knexfile and migrations to src subdirectory
